+++ src/NamedThingVector.cpp
@@ -67,7 +67,7 @@ void* NamedThingVector::getElement(int index) {
 }
 
 String* NamedThingVector::getName(void* namedThing) {
-	return (String*)((uint32_t)namedThing + stringOffset);
+	return (String*)((uintptr_t)namedThing + stringOffset);
 }
 
 // Returns error code
